Create modules fot js files, Add live demo link, Update ReadMe
