+++ package.json
@@ -1,31 +1,24 @@
 {
-  "name": "to-do-list",
+  "name": "webpack-demo",
   "version": "1.0.0",
-  "description": "\"To-do list\" is a tool that helps to organize your day. It simply lists the things that you need to do and allows you to mark them as complete.",
-  "main": "index.js",
+  "description": "",
+  "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "webpack serve --open",
     "build": "webpack",
-    "start": "webpack-dev-server --hot",
-    "dev": "webpack serve"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/felixDev22/To-Do-List.git"
+    "deploy": "gh-pages -d dist"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/felixDev22/To-Do-List/issues"
-  },
-  "homepage": "https://github.com/felixDev22/To-Do-List#readme",
   "devDependencies": {
     "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.2",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.26.0",
+    "gh-pages": "^4.0.0",
     "hint": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
@@ -35,8 +28,7 @@
     "stylelint-scss": "^3.21.0",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.11.1",
-    "webpack-serve": "^4.0.0"
+    "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
     "lodash": "^4.17.21"
