+++ src/index.js
@@ -1,80 +1,20 @@
 import './style.css';
 
-const listItems = document.querySelector('.task-list');
-const addTask = document.querySelector('#new-task');
-const addBtn = document.getElementById('plus');
+import { listItems, addTask, addBtn } from './modules/variables.js';
+import {
+  addNewList,
+  generateList,
+  pushToLocal,
+  showList,
+  removeList,
+  checkCompleted,
+} from './modules/display.js';
+import removeAllCompleted from './modules/clearCompleted.js';
 
-let taskArr = [];
-
-const addNewList = () => {
-  const task = {};
-  task.index = taskArr.length + 1;
-  task.description = addTask.value;
-  task.completed = false;
-  taskArr.push(task);
-};
-
-// Save To Local Directory
-const pushToLocal = () => {
-  localStorage.setItem('taskArr', JSON.stringify(taskArr));
-};
-
-// To render the task list
-const generateList = () => {
-  taskArr.forEach((task) => {
-    listItems.innerHTML += `
-        <li class="task" >
-          <div class='item' >
-          <input data-action="checkbox" type="Checkbox" id='1' data-id="${task.index}" />
-          <p class="description">"${task.description}"</p>
-          </div class='item'>
-          
-          <div id="edit" class="hide">
-         <i class="fa-solid fa-ellipsis-vertical"></i>
-          </div>
-         
-          <i data-action="delete" class="fa-regular fa-trash-can" id="removeBtn"></i>
-         
-        </li>`;
-    addTask.value = '';
-  });
-};
-
-const showList = () => {
-  if (localStorage.getItem('taskArr')) {
-    taskArr = JSON.parse(localStorage.getItem('taskArr'));
-  }
-  generateList();
-};
-
-// Delete Button to remove task
-const removeList = (index) => {
-  taskArr.splice(index, 1);
-  addNewList();
-  generateList();
-  pushToLocal();
-};
-
-// Checkbox
-const boxChecked = (box) => {
-  if (box.checked) {
-    return true;
-  }
-  return false;
-};
-
-// Completed task list
-const checkCompleted = (buttonId, box) => {
-  box.nextElementSibling.classList.toggle('mark-completed');
-  taskArr[buttonId].completed = boxChecked(box);
-  pushToLocal(taskArr);
-  if (taskArr[buttonId].completed === true) {
-    box.checked = true;
-    box.nextElementSibling.classList.add('mark-completed');
-  }
-};
+import boxChecked from './modules/completed.js';
 
 // on Load
+
 window.addEventListener('load', () => {
   showList();
 });
@@ -111,6 +51,7 @@ listItems.addEventListener('click', (e) => {
 });
 
 // Highlight Completed task
+
 listItems.addEventListener('change', (e) => {
   const { target } = e;
   const parentElement = target.parentNode;
@@ -124,16 +65,6 @@ listItems.addEventListener('change', (e) => {
     boxChecked(target);
   }
 });
-
-// remove all completed
-const removeAllCompleted = () => {
-  const newTasks = taskArr.filter((taskArr) => !taskArr.completed);
-  for (let i = 0; i < newTasks.length; i += 1) {
-    newTasks[i].index = i + 1;
-  }
-  taskArr.length = 0;
-  newTasks.forEach((newTask) => taskArr.push(newTask));
-};
 
 // Clear completed tasks
 const clearButton = document.querySelector('.clear');
