+++ src/modules/completed.js
@@ -1,0 +1,7 @@
+const boxChecked = (box) => {
+  if (box.checked) {
+    return true;
+  }
+  return false;
+};
+export default boxChecked;
